test(rewards): scope reward fixtures to a rewards conftest
